removed unnecessary use of functors
