+++ sazanov.maxim/T3/CommandFacade.hpp
@@ -1,5 +1,5 @@
-#ifndef COMMAND_CENTER_HPP
-#define COMMAND_CENTER_HPP
+#ifndef COMMAND_FACADE_HPP
+#define COMMAND_FACADE_HPP
 
 #include <string>
 #include <unordered_map>
@@ -8,25 +8,25 @@
 #include <exception>
 #include <iostream>
 #include "Polygon.hpp"
-#include "CommandFunctors.hpp"
+#include "Commands.hpp"
 
 namespace sazanov
 {
   class CommandFacade
   {
-    using AccumulateFunctor = std::function< double(double, const Polygon&) >;
-    using AreaSubCommands = std::unordered_map< std::string, AccumulateFunctor >;
+    using AccumulatePredicate = std::function< double(double, const Polygon&) >;
+    using AreaSubCommands = std::unordered_map< std::string, AccumulatePredicate >;
 
     using Comparator = std::function< bool(const Polygon&, const Polygon&) >;
-    using OutputValueFunctor = std::function< void(const Polygon&, std::ostream& out) >;
-    using MaxMinSubCommands = std::unordered_map< std::string, std::pair<Comparator, OutputValueFunctor > >;
+    using OutputValuePredicate = std::function< void(const Polygon&, std::ostream& out) >;
+    using MaxMinSubCommands = std::unordered_map< std::string, std::pair< Comparator, OutputValuePredicate > >;
 
-    using CountFunctor = std::function< bool(const Polygon&) >;
-    using CountSubCommands = std::unordered_map<std::string, CountFunctor>;
+    using CountPredicate = std::function< bool(const Polygon&) >;
+    using CountSubCommands = std::unordered_map< std::string, CountPredicate >;
 
   public:
     explicit CommandFacade(const std::vector< Polygon >& polygons);
-    void nextCommand(std::istream& in, std::ostream& out);
+    void nextCommand(std::istream& in, std::ostream& out) const;
 
   private:
     const std::vector< Polygon >& polygons_;
@@ -34,10 +34,10 @@ namespace sazanov
     using CommandFunctor = std::function< void(std::istream& in, std::ostream& out) >;
     std::unordered_map< std::string, CommandFunctor > commands_;
 
-    AreaSubCommands getAreaSubCommands();
-    MaxMinSubCommands getMaxMinSubCommands();
-    CountSubCommands getCountSubcommands();
-    std::unordered_map< std::string, bool > getEmptyVectorSupportInfo();
+    AreaSubCommands getAreaSubCommands() const;
+    MaxMinSubCommands getMaxMinSubCommands() const;
+    CountSubCommands getCountSubcommands() const;
+    std::unordered_map< std::string, bool > getEmptyVectorSupportInfo() const;
   };
 }
